refactor: set instance of controller in application
