+++ src/controllers/board.js
@@ -1,7 +1,14 @@
 
-import {render, replace, RenderPosition} from '../utils.js';
+import {render, replace, remove, RenderPosition} from '../utils.js';
 import TaskComponent from '../components/task.js';
 import TaskEditComponent from '../components/task-edit.js';
+import TasksComponent from '../components/tasks.js';
+import NoTasksComponent from '../components/no-tasks.js';
+import SortComponent from '../components/sorting.js';
+import LoadMoreButtonComponent from '../components/load-more-button.js';
+
+const SHOWING_TASKS_COUNT_ON_START = 8;
+const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
 
 const renderTask = (taskListElement, task) => {
   const onEscKeyDown = (evt) => {
@@ -36,7 +43,49 @@ const renderTask = (taskListElement, task) => {
 };
 
 export default class BoardController {
+  constructor(container) {
+    this._container = container;
 
+    this._noTasksComponent = new NoTasksComponent();
+    this._sortComponent = new SortComponent();
+    this._tasksComponent = new TasksComponent();
+    this._loadMoreButtonComponent = new LoadMoreButtonComponent();
+  }
+
+  render(tasks) {
+    const container = this._container.getElement();
+    const isAllTasksArchived = tasks.every((task) => task.isArchive);
+
+    if (isAllTasksArchived) {
+      render(container, this._noTasksComponent, RenderPosition.BEFOREEND);
+      return;
+    }
+
+    render(container, this._sortComponent, RenderPosition.BEFOREEND);
+    render(container, this._tasksComponent, RenderPosition.BEFOREEND);
+
+    const taskListElement = this._tasksComponent.getElement();
+
+    let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
+    tasks.slice(0, showingTasksCount)
+      .forEach((task) => {
+        renderTask(taskListElement, task);
+      });
+
+    render(container, this._loadMoreButtonComponent, RenderPosition.BEFOREEND);
+
+    this._loadMoreButtonComponent.setClickHandler(() => {
+      const prevTasksCount = showingTasksCount;
+      showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
+
+      tasks.slice(prevTasksCount, showingTasksCount)
+        .forEach((task) => renderTask(taskListElement, task));
+
+      if (showingTasksCount >= tasks.length) {
+        remove(this._loadMoreButtonComponent);
+      }
+    });
+  }
 }
 
 export {renderTask};
